+++ packages/hooks/src/context.tsx
@@ -8,7 +8,7 @@ import React, {
   useMemo,
   useState,
 } from 'react'
-import type { Config, Sdk } from './graphql'
+import type { Sdk } from './graphql'
 import { getSdk } from './graphql'
 
 gql`
@@ -28,7 +28,6 @@ export type LiteflowContext = {
   resetAuthenticationToken: () => void
   currentAddress: string | null
   sdk: Sdk
-  config: () => Promise<Config>
 }
 
 export type LiteflowProviderProps = {
@@ -44,7 +43,6 @@ export const LiteflowContext = createContext<LiteflowContext>({
   },
   currentAddress: null,
   sdk: {} as Sdk,
-  config: () => ({} as Promise<Config>),
 })
 
 export function LiteflowProvider({
@@ -54,13 +52,6 @@ export function LiteflowProvider({
   const [authenticationToken, setAuthenticationToken] = useState<string>()
   const client = useMemo(() => new GraphQLClient(endpoint), [endpoint])
   const sdk = useMemo(() => getSdk(client), [client])
-  const [configPromise, setConfigPromise] = useState<Promise<Config>>()
-  const config = useCallback(() => {
-    if (configPromise) return configPromise
-    const promise = sdk.GetConfig().then(({ config }) => config)
-    setConfigPromise(promise)
-    return promise
-  }, [sdk, configPromise, setConfigPromise])
   const currentAddress = useMemo(() => {
     if (!authenticationToken) return null
     const res = decode<JwtPayload & { address: string }>(authenticationToken)
@@ -88,7 +79,6 @@ export function LiteflowProvider({
         setAuthenticationToken,
         currentAddress,
         sdk,
-        config,
       }}
     >
       {children}