+++ packages/hooks/CHANGELOG.md
@@ -4,8 +4,8 @@
 
 #### Breaking Changes
 
-- Update `useConfig` return parameter from `Promise<Config>` to `() => Promise<Config>` [#]()
-  - Make sure to update your codebase with `const getConfig = useConfig(); ... await getConfig()`
+- Update `useConfig` return parameter from `Promise<Config>` to `{ data?: Config, error?: Error }` [#120](https://github.com/liteflow-labs/liteflow-js/pull/120)
+  - Make sure to update your codebase with `const { data } = useConfig();`
 
 #### Added
 
@@ -17,7 +17,7 @@
 
 #### Fixed
 
-- Fix issue on initial render on SSR due to pending promise [#]()
+- Fix issue on initial render on SSR due to pending promise [#120](https://github.com/liteflow-labs/liteflow-js/pull/120)
 - Fix imports order by updating `prettier` and `prettier-plugin-organize-imports` [#114](https://github.com/liteflow-labs/liteflow-js/pull/114)
 
 #### Security